Cleanup and deps update
