+++ LRReader.Shared/Extensions/Extensions.cs
@@ -1,31 +1,40 @@
 ﻿using System;
 
-namespace LRReader.Shared.Extensions
+namespace LRReader.Shared.Extensions;
+
+public static class StringExtensions
 {
-	public static class StringExtensions
+	public static string AsFormat(this string format, params object[] args) => string.Format(format, args);
+
+	public static string UpperFirstLetter(this string str)
 	{
-		public static string AsFormat(this string format, params object[] args) => string.Format(format, args);
+		if (str.Length == 0)
+			return "";
+		else if (str.Length == 1)
+			return char.ToUpper(str[0]).ToString();
+		else
+			return char.ToUpper(str[0]) + str.Substring(1);
 	}
+}
 
-	public static class CompareExtensions
+public static class CompareExtensions
+{
+	public static T Clamp<T>(this T val, T min, T max) where T : IComparable<T>
 	{
-		public static T Clamp<T>(this T val, T min, T max) where T : IComparable<T>
-		{
-			if (val.CompareTo(min) < 0) return min;
-			else if (val.CompareTo(max) > 0) return max;
-			else return val;
-		}
+		if (val.CompareTo(min) < 0) return min;
+		else if (val.CompareTo(max) > 0) return max;
+		else return val;
 	}
+}
 
-	public class GridViewExtParameter
+public class GridViewExtParameter
+{
+	public bool Ctrl { get; }
+	public object Item { get; }
+
+	public GridViewExtParameter(bool ctrl, object item)
 	{
-		public bool Ctrl { get; }
-		public object Item { get; }
-
-		public GridViewExtParameter(bool ctrl, object item)
-		{
-			Ctrl = ctrl;
-			Item = item;
-		}
+		Ctrl = ctrl;
+		Item = item;
 	}
 }
