+++ LRReader.Shared/Generic.cs
@@ -1,14 +1,13 @@
 ﻿using System.Threading.Tasks;
 
-namespace LRReader.Shared
-{
-	public delegate Task AsyncAction<T>(T obj);
+namespace LRReader.Shared;
 
-	public static class AsyncActionExtensions
+public delegate Task AsyncAction<T>(T obj);
+
+public static class AsyncActionExtensions
+{
+	public static Task InvokeAsync<T>(this AsyncAction<T>? action, T obj)
 	{
-		public static Task InvokeAsync<T>(this AsyncAction<T>? action, T obj)
-		{
-			return action?.Invoke(obj) ?? Task.CompletedTask;
-		}
+		return action?.Invoke(obj) ?? Task.CompletedTask;
 	}
 }
