+++ LRReader.Shared/LRReader.Shared.csproj
@@ -15,8 +15,8 @@
     <PackageReference Include="Microsoft.AppCenter.Crashes" Version="5.0.3" Condition="'$(TargetFramework)' == 'uap10.0.17763'" />
     <PackageReference Include="Caching.dll" Version="2.0.0.1" />
     <PackageReference Include="KeyedSemaphores" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="7.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="NReco.Logging.File" Version="1.1.7" />
     <PackageReference Include="Nullable" Version="1.3.1">
