+++ LRReader.Shared/Models/Main/Category.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using LRReader.Shared.Converters;
 using Newtonsoft.Json;
 
 namespace LRReader.Shared.Models.Main
