+++ LRReader.Shared/Models/Main/Generic.cs
@@ -1,4 +1,5 @@
 ﻿using System.Net;
+using LRReader.Shared.Converters;
 using Newtonsoft.Json;
 
 namespace LRReader.Shared.Models.Main
