+++ LRReader.Shared/Models/Main/Server.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using LRReader.Shared.Converters;
 using Newtonsoft.Json;
 
 namespace LRReader.Shared.Models.Main
