+++ LRReader.Shared/Models/Main/Shinobu.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using LRReader.Shared.Converters;
+using Newtonsoft.Json;
 
 namespace LRReader.Shared.Models.Main
 {
