+++ LRReader.UWP.Installer/LRReader.UWP.Installer.csproj
@@ -34,7 +34,7 @@
     </PackageReference>
     <PackageReference Include="Microsoft.Windows.SDK.Contracts" Version="10.0.22621.2428" />
     <PackageReference Include="WPF-UI">
-      <Version>3.0.0-preview.8</Version>
+      <Version>3.0.0-preview.9</Version>
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
