+++ LRReader.UWP/Converters/Converters.cs
