+++ LRReader.UWP/LRReader.UWP.csproj
@@ -155,7 +155,7 @@
     <Compile Include="Services\SettingsStorage.cs" />
     <Compile Include="Init.cs" />
     <Compile Include="Extensions\Extensions.cs" />
-    <Compile Include="Internal\Converters.cs" />
+    <Compile Include="Converters\Converters.cs" />
     <Compile Include="Services\ImageProcessing.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Secrets.cs" />
