+++ LRReader.UWP/Views/Tabs/Content/ArchiveTabContent.xaml
@@ -56,6 +56,25 @@
 										Text="{x:Bind Data.Archive.title, Mode=OneWay}"
 										TextWrapping="WrapWholeWords" />
 								</Grid>
+								<!--
+								<StackPanel Orientation="Horizontal" Spacing="4" Margin="0,0,0,8">
+									<Border Padding="4,2" BorderThickness="1" CornerRadius="{ThemeResource ControlCornerRadius}" BorderBrush="{ThemeResource ControlElevationBorderBrush}" BackgroundSizing="InnerBorderEdge" Background="{ThemeResource ControlFillColorDefault}" >
+										<TextBlock>
+											Pages: <Run Text="{x:Bind Data.Archive.pagecount, Mode=OneWay}"/>
+										</TextBlock>
+									</Border>
+									<Border Padding="4,2" BorderThickness="1" CornerRadius="{ThemeResource ControlCornerRadius}" BorderBrush="{ThemeResource ControlElevationBorderBrush}" BackgroundSizing="InnerBorderEdge" Background="{ThemeResource ControlFillColorDefault}" >
+										<TextBlock>
+											Last read: <Run Text="{x:Bind util2:Util.UnixTimeToDateTime(Data.Archive.lastreadtime), Mode=OneWay}"/>
+										</TextBlock>
+									</Border>
+									<Border Padding="4,2" BorderThickness="1" CornerRadius="{ThemeResource ControlCornerRadius}" BorderBrush="{ThemeResource ControlElevationBorderBrush}" BackgroundSizing="InnerBorderEdge" Background="{ThemeResource ControlFillColorDefault}" >
+										<TextBlock>
+											File type: <Run Text="{x:Bind Data.Archive.extension, Mode=OneWay}"/>
+										</TextBlock>
+									</Border>
+								</StackPanel>
+								-->
 								<controls:ArchiveTags Margin="0,0,0,8" ItemClickCommand="{x:Bind Data.TagClickCommand}" ItemsSource="{x:Bind Data.Archive.TagsGroups, Mode=OneWay}" />
 								<StackPanel Orientation="Horizontal">
 									<ToggleButton x:Uid="/Tabs/Archive/Bookmark" IsChecked="{x:Bind Data.Bookmarked, Mode=TwoWay}" IsEnabled="{x:Bind Data.ControlsEnabled, Mode=OneWay}" />
